% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_expression}
\alias{add_expression}
\title{attach expression values from user-provided expression vector to graph.}
\usage{
add_expression(exp, g)
}
\arguments{
\item{exp}{expression vector - assumed to be a named vector where the values are expression and the names are the gene name}

\item{g}{igraph object - will be filtered so that only nodes found in both exp and g are kept}
}
\value{
subgraph of g containing only shared keys with exp and with expression attached.
}
\description{
attach expression values from user-provided expression vector to graph.
}
