% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_value}
\alias{add_value}
\title{Attach a generic user-provided value to graph}
\usage{
add_value(val, g, val_name = "value")
}
\arguments{
\item{val}{named numeric vector where the names correspond to vertices in g}

\item{g}{igraph object - will be filtered so that only nodes found in both exp and g are kept}

\item{val_name}{str key for val}
}
\value{
subgraph of g containing only shared keys with val and val attached
}
\description{
Attach a generic user-provided value to graph
}
