% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_gene_symbol}
\alias{as_gene_symbol}
\title{Convert from ensembl.gene to gene.symbol}
\usage{
as_gene_symbol(x, edb = EnsDb.Hsapiens.v79::EnsDb.Hsapiens.v79)
}
\arguments{
\item{x}{vector of ensemble.gene ids, ensemble.peptide ids, ensemble.transcript
ids or entrez gene ids}

\item{edb}{ensemble database object}
}
\value{
vector of gene symbols
}
\description{
Convert from ensembl.gene to gene.symbol
}
