% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_null.R
\name{bootstrap_null}
\alias{bootstrap_null}
\title{Bootstrap null distribution for significance testing}
\usage{
bootstrap_null(
  ppi = "biogrid",
  n = 1000,
  seed_proteins,
  gamma,
  eps = 1e-10,
  tmax = 1000,
  norm = TRUE
)
}
\description{
Bootstrap null distribution for significance testing
}
