% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_np.R
\name{calc_np}
\alias{calc_np}
\title{calculate network potential for one node.}
\usage{
calc_np(c_i, c_j)
}
\arguments{
\item{c_i}{expression for a given node.}

\item{c_j}{vector of expressions for each neighbor of c_i}
}
\description{
calculate network potential for one node.
}
