% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_np.R
\name{calc_np_all_legacy}
\alias{calc_np_all_legacy}
\title{function to calculate the network potential for each protein in a user-provided vector}
\usage{
calc_np_all_legacy(
  exp,
  g,
  v = as.character(names(igraph::V(g))),
  neighbors = NULL
)
}
\arguments{
\item{exp}{expression vector - assumed to be a named vector where the values are expression and the names are the gene name}

\item{g}{igraph object - will be filtered so that only nodes found in both exp and g are kept}

\item{v}{character vector of nodes over which to calculate network potential.}

\item{neighbors}{named list containing the neighbors for each node of graph g. If not provided,
it will be computed}
}
\value{
dataframe containing network potential for each of the inputed gene names.
}
\description{
Mostly just used to help debug the CPP version - not exported
}
