% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{check_crosstalk}
\alias{check_crosstalk}
\title{Check to make sure incoming object is a valid crosstalk df.}
\usage{
check_crosstalk(crosstalk_df)
}
\arguments{
\item{crosstalk_df}{a dataframe containing the results of \code{compute_crosstalk}}
}
\value{
message if not correct object type, null otherwise
}
\description{
This function is a helper function for \code{plot_ct} that verifies the input is a valid output of compute_crosstalk
}
