% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_null.R
\name{combine_null}
\alias{combine_null}
\title{.combine function for compute_null foreach looping structure}
\usage{
combine_null(x)
}
\arguments{
\item{x}{aggregated data structure}
}
\value{
data.frame
}
\description{
.combine function for compute_null foreach looping structure
}
