% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_crosstalk.R
\name{compute_crosstalk}
\alias{compute_crosstalk}
\title{Identify proteins with a statistically significant relationship to user-provided seeds.}
\usage{
compute_crosstalk(
  seed_proteins,
  ppi = "stringdb",
  n = 10000,
  gamma = 0.6,
  eps = 1e-10,
  tmax = 1000,
  norm = TRUE,
  set_seed,
  cache,
  seed_name = NULL,
  ncores = 1
)
}
\arguments{
\item{seed_proteins}{user defined seed proteins}

\item{ppi}{specifies which protein-protein interaction network will be used - currently only "stringdb" is supported}

\item{n}{number of random walks with repeats}

\item{gamma}{restart probability}

\item{eps}{maximum allowed difference between the computed probabilities at the steady state}

\item{tmax}{the maximum number of iterations for the RWR}

\item{norm}{if True, w is normalized by dividing each value by the column sum.}

\item{set_seed}{integer to set random number seed - for reproducibility}

\item{cache}{A filepath to a folder downloaded files should be stored, inherits from user-available functions}

\item{seed_name}{Name to give the cached null distribution - must be a character string}

\item{ncores}{= 1}
}
\description{
Identify proteins with a statistically significant relationship to user-provided seeds.
}
