% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_inputtype}
\alias{detect_inputtype}
\title{Determine which format of gene is used to specify by user-defined seed proteins}
\usage{
detect_inputtype(x)
}
\arguments{
\item{x}{vector of gene symbols}
}
\value{
"gene_symbol", "entrez_id", "ensemble_id" or "other"
}
\description{
Determine which format of gene is used to specify by user-defined seed proteins
}
