% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_null.R
\name{dist_calc}
\alias{dist_calc}
\title{Internal function that computes the mean/stdev for each gene from a wide-format data frame.}
\usage{
dist_calc(df, ncores, seed_proteins)
}
\arguments{
\item{df}{: numeric vector}
}
\value{
a 3-column dataframe (gene, )
}
\description{
This function is called by the high-level function "bootstrap_null".
}
