% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensembl_convert}
\alias{ensembl_convert}
\title{Convert from ensembl.gene to gene.symbol}
\usage{
ensembl_convert(x, edb = EnsDb.Hsapiens.v79::EnsDb.Hsapiens.v79)
}
\arguments{
\item{x}{vector of ensemble.gene ids}

\item{edb}{ensemble database object}
}
\value{
vector of gene.symbol
}
\description{
Convert from ensembl.gene to gene.symbol
}
