% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensembl_type}
\alias{ensembl_type}
\title{Determine if ensembl id is a Protein, gene, or transcript_id}
\usage{
ensembl_type(x)
}
\arguments{
\item{x}{vector or single gene symbol}
}
\value{
character: "PROTEINID", "GENEID", "TRANSCRIPTID"
}
\description{
Determine if ensembl id is a Protein, gene, or transcript_id
}
