% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fcalc_np_all}
\alias{fcalc_np_all}
\title{Function to calculate the network potential for vertices v}
\usage{
fcalc_np_all(neighbors, vertices, v, exp)
}
\arguments{
\item{neighbors}{list of neighbors for every node in the graph, type Rcpp::list}

\item{vertices}{node list for graph, type Rcpp::StringVector}

\item{v}{list of nodes for which we plan to calculate network potential}

\item{exp}{named vector of expression for each node in vertices}
}
\description{
Function to calculate the network potential for vertices v
}
