% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_null.R
\name{final_dist_calc}
\alias{final_dist_calc}
\title{Internal function that computes the mean/stdev for each gene from a wide-format data frame.}
\usage{
final_dist_calc(df_list)
}
\arguments{
\item{df_list}{: list of dataframes from foreach loop in bootstrap_null}
}
\value{
a dataframe
}
\description{
This function is called by the high-level function "bootstrap_null".
}
