% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_neighbors}
\alias{get_neighbors}
\title{function to get graph neighbors (along with their expression values) for a given gene in a given network g}
\usage{
get_neighbors(gene, g)
}
\arguments{
\item{gene}{gene to grab neighbors from.}

\item{g}{igraph object - will be filtered so that only nodes found in both exp and g are kept}
}
\value{
named numeric vector.
}
\description{
just a wrapper around \code{\link[igraph:neighbors]{igraph::neighbors()}} for convenience
}
