% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_null.R
\name{get_topn}
\alias{get_topn}
\title{Helper function for compute_null_dnp - returns the top n genes by dnp for each sample}
\usage{
get_topn(df, n_genes)
}
\arguments{
\item{df}{output of \code{\link[=compute_dnp]{compute_dnp()}}}

\item{n_genes}{integer describing number of genes per sample that we will compute the null distribution for}
}
\description{
Helper function for compute_null_dnp - returns the top n genes by dnp for each sample
}
