% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{gfilter.igraph_method}
\alias{gfilter.igraph_method}
\title{Method to filter graph based on an igraph method that scores verticies.}
\usage{
gfilter.igraph_method(g, ppi = TRUE, method, n = 500, desc, ...)
}
\arguments{
\item{g}{igraph object}

\item{ppi}{bool - should we use a ppi from online repository?}

\item{method}{str, supported methods can be found by calling \code{\link[=supported_filter_methods]{supported_filter_methods()}}}

\item{n}{int - number of nodes to include in the returned subgraph}

\item{desc}{bool - do we want the top or bottom examples of the provided metri}

\item{...}{additional params passed to \code{\link[=load_ppi]{load_ppi()}}}
}
\value{
igraph
}
\description{
Tested methods include: \code{\link[igraph:betweenness]{igraph::betweenness()}},
}
