% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{gfilter.value}
\alias{gfilter.value}
\title{Method to filter graph based on user provided value}
\usage{
gfilter.value(g, val, ppi = TRUE, n = 500, desc, ...)
}
\arguments{
\item{g}{igraph object}

\item{val}{named numeric vector - some measure of node state (i.e. gene expression in the case of a PPI)}

\item{ppi}{bool - should we use a ppi from online repository?}

\item{n}{int - number of nodes to include in the returned subgraph}

\item{desc}{bool - do we want the top or bottom examples of the provided metri}

\item{...}{additional params passed to \code{\link[=load_ppi]{load_ppi()}}}
}
\value{
igraph
}
\description{
Method to filter graph based on user provided value
}
