% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_ensembl}
\alias{is_ensembl}
\title{Determine if a character vector contains ensembl gene_ids}
\usage{
is_ensembl(x)
}
\arguments{
\item{x}{vector or single gene symbol}
}
\value{
logical
}
\description{
Determine if a character vector contains ensembl gene_ids
}
