% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_entrez}
\alias{is_entrez}
\title{Determine if a character vector contains entrez gene_ids}
\usage{
is_entrez(x)
}
\arguments{
\item{x}{vector or single gene symbol}
}
\value{
logical
}
\description{
Determine if a character vector contains entrez gene_ids
}
