% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_ingest.R
\name{load_ppi}
\alias{load_ppi}
\title{Helper function to load requested PPI w/ parameters}
\usage{
load_ppi(
  cache = NULL,
  union = FALSE,
  intersection = FALSE,
  species = "9606",
  min_score = 0,
  ppi = "stringdb"
)
}
\arguments{
\item{cache}{A filepath to a folder downloaded files should be stored}

\item{union}{bool}

\item{intersection}{bool}

\item{species}{species code either using latin species name or taxon id}

\item{min_score}{minimum connectivity score for each edge in the network.}

\item{ppi}{str}
}
\value{
igraph object
}
\description{
Helper function to load requested PPI w/ parameters
}
