% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_null.R
\name{match_seeds}
\alias{match_seeds}
\title{Identify random sets of seeds with similar degree distribution to parent seed proteins}
\usage{
match_seeds(g, seed_proteins, n, set_seed = NULL)
}
\arguments{
\item{g}{igraph object representing the network under study. specified by "ppi" in bootstrap_null}

\item{seed_proteins}{user defined seed proteins}

\item{n}{number of random walks with repeats to create null distribution}

\item{set_seed}{integer to set random number seed - for reproducibility}
}
\value{
list of character vectors: randomly generated seed proteins with a similar
degree distribution to parent seed proteins
}
\description{
This function will generate n character vectors of seeds to be passed to sparseRWR
as part of the construction of a boostrapped null distribution for significance testing.
}
