% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_suppression.R
\name{node_repression}
\alias{node_repression}
\title{Function to eliminate a node from a network g and calculate the change in some measure of network state}
\usage{
node_repression(
  g,
  v_rm,
  exp,
  state_function = calc_np_all,
  neighbors_only = TRUE,
  ...
)
}
\arguments{
\item{g}{igraph network object}

\item{v_rm}{index of vertices to remove}

\item{exp}{expression vector for nodes in graph g}

\item{state_function}{function to use to calculate network state before and after node_repression}

\item{neighbors_only}{logical designating whether state function should be calculated for all nodes or just neighbors}

\item{...}{additional parameters passed to state function.}
}
\description{
this function is still under development.
}
