% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalkRepeats.R
\name{norm1}
\alias{norm1}
\title{Function that computes the norm 1 of a numeric vector}
\usage{
norm1(x)
}
\arguments{
\item{x}{: numeric vector}
}
\value{
a single real value (the norm1 of the input vector)
}
\description{
This function is reproduced from the source code of the RANKS package (not exported).
}
