% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalkRepeats.R
\name{norm_colsum}
\alias{norm_colsum}
\title{Function to normalize adjacency matrix by dividing each value by the colsum.}
\usage{
norm_colsum(w)
}
\arguments{
\item{w}{The adjacency matrix of a given graph in sparse format - dgCMatrix}
}
\description{
Function to normalize adjacency matrix by dividing each value by the colsum.
}
