% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_ingest.R
\name{ppi_intersection}
\alias{ppi_intersection}
\title{Function to allow users to choose the intersection of stringdb and biogrid}
\usage{
ppi_intersection(cache = NULL, min_score = 0, edb = "default")
}
\arguments{
\item{cache}{A filepath to a folder downloaded files should be stored}

\item{min_score}{minimum connectivity score for each edge in the network.}

\item{edb}{ensemble database object}
}
\value{
igraph object corresponding to PPI following intersection
}
\description{
Function to allow users to choose the intersection of stringdb and biogrid
}
