% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_ingest.R
\name{prep_biogrid}
\alias{prep_biogrid}
\title{Prepare biogrid for use in analyses}
\usage{
prep_biogrid(cache = NULL)
}
\arguments{
\item{cache}{A filepath to a folder downloaded files should be stored}
}
\value{
igraph object built from the adjacency matrix downloaded from thebiogrid.org.
}
\description{
Prepare biogrid for use in analyses
}
