% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_ingest.R
\name{prep_stringdb}
\alias{prep_stringdb}
\title{Prepare Stringdb for use in analyses}
\usage{
prep_stringdb(cache = NULL, edb = "default", min_score = NULL)
}
\arguments{
\item{cache}{A filepath to a folder downloaded files should be stored, inherits from user-available functions}

\item{edb}{ensemble database object}

\item{min_score}{minimum connectivity score for each edge in the network.}
}
\value{
list containing Adjacency matrix from stringdb dataset and igraph object built from the adjacency matrix.
}
\description{
Prepare Stringdb for use in analyses
}
