% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_ingest.R
\name{setup_init}
\alias{setup_init}
\title{Helper function for first-time use of crosstalkr package}
\usage{
setup_init(cache = NULL, min_score)
}
\arguments{
\item{cache}{A filepath to a folder downloaded files should be stored, inherits from user-available functions}

\item{min_score}{minimum connectivity score for each edge in the network.}
}
\value{
directory on users computer containing the different adjacency matrices for future use.
}
\description{
Helper function for first-time use of crosstalkr package
}
