% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalkRepeats.R
\name{sparseRWR}
\alias{sparseRWR}
\title{Perform random walk with repeats on a sparse matrix}
\usage{
sparseRWR(w, seed_proteins, gamma = 0.6, eps = 1e-10, tmax = 1000, norm = TRUE)
}
\arguments{
\item{w}{The adjacency matrix of a given graph in sparse format - dgCMatrix}

\item{seed_proteins}{user defined seed proteins}

\item{gamma}{restart probability}

\item{eps}{maximum allowed difference between the computed probabilities at the steady state}

\item{tmax}{the maximum number of iterations for the RWR}

\item{norm}{if True, w is normalized by dividing each value by the column sum.}
}
\description{
This function borrows heavily from the RWR function in the RANKS package (cite here)
}
