% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{supported_filter_methods}
\alias{supported_filter_methods}
\title{return all supported filter methods with their descriptions}
\usage{
supported_filter_methods()
}
\value{
data.frame
}
\description{
return all supported filter methods with their descriptions
}
