% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_ingest.R
\name{supported_species}
\alias{supported_species}
\title{returns a dataframe with information on supported species}
\usage{
supported_species()
}
\value{
dataframe
}
\description{
returns a dataframe with information on supported species
}
