% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_pipeline.R
\name{tidy_expression}
\alias{tidy_expression}
\title{helper function to convert expression matrix to tidy dataframe (if not already)}
\usage{
tidy_expression(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
data.frame
}
\description{
helper function to convert expression matrix to tidy dataframe (if not already)
}
