% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_ingest.R
\name{to_taxon_id}
\alias{to_taxon_id}
\title{helper to convert user-inputs to ncbi reference taxonomy.}
\usage{
to_taxon_id(species)
}
\arguments{
\item{species}{user-inputted species}
}
\value{
string corresponding to taxon id
}
\description{
helper to convert user-inputs to ncbi reference taxonomy.
}
