% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_null.R
\name{combine_null}
\alias{combine_null}
\title{.combine function for compute_null foreach looping structure}
\usage{
combine_null(x)
}
\arguments{
\item{x}{aggregated data structure}

\item{y}{task returned by inner foreach loop of compute_null}

\item{...}{arbitrary number of tasks returned by the inner foreach loop of
compute_null_dnp}
}
\value{
data.frame
}
\description{
.combine function for compute_null foreach looping structure
}
