% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_null.R
\name{dist_calc}
\alias{dist_calc}
\title{Internal function that computes the mean/stdev for each gene from a wide-format data frame.}
\usage{
dist_calc(df, seed_proteins)
}
\arguments{
\item{df}{: numeric vector}

\item{seed_proteins}{user defined seed proteins}
}
\value{
a data frame containing summary statistics for the computed null distribution
}
\description{
This function is called by the high-level function "bootstrap_null".
Not expected to be used by end-users - we only export it so that environments
inside foreach loops can find it.
}
