% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_null.R
\name{final_combine}
\alias{final_combine}
\title{final .combine function to run in compute_null_dnp foreach looping structure}
\usage{
final_combine(x)
}
\arguments{
\item{x}{aggregated info}

\item{...}{arbitrary number of tasks returned by the inner foreach loop of
compute_null_dnp}
}
\value{
data.frame
}
\description{
final .combine function to run in compute_null_dnp foreach looping structure
}
