% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{gfilter.ct}
\alias{gfilter.ct}
\title{method to filter the graph based on parameters passed to compute_crosstalk}
\usage{
gfilter.ct(seeds, ...)
}
\arguments{
\item{...}{additional arguments passed to \code{\link[=compute_crosstalk]{compute_crosstalk()}}}
}
\value{
igraph object
}
\description{
method to filter the graph based on parameters passed to compute_crosstalk
}
