% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{gfilter.ct}
\alias{gfilter.ct}
\title{Method to filter the graph based on parameters passed to compute_crosstalk}
\usage{
gfilter.ct(seeds, ...)
}
\arguments{
\item{seeds}{vector (str or numeric) user provided vertex ids to use as seeds in the random walk with restarts}

\item{...}{additional arguments passed to \code{\link[=compute_crosstalk]{compute_crosstalk()}}}
}
\value{
igraph object
}
\description{
Method to filter the graph based on parameters passed to compute_crosstalk
}
