% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{gfilter.igraph_method}
\alias{gfilter.igraph_method}
\title{Method to filter graph based on any igraph method that scores verticies.}
\usage{
gfilter.igraph_method(g, use_ppi = TRUE, method, n = 500, desc, val_name, ...)
}
\arguments{
\item{g}{igraph object}

\item{use_ppi}{bool - should we use a ppi from online repository?}

\item{method}{str, supported methods can be found by calling \code{\link[=supported_filter_methods]{supported_filter_methods()}}}

\item{n}{int - number of nodes to include in the returned subgraph}

\item{desc}{bool - do we want the top or bottom examples of the provided metric}

\item{val_name}{str}

\item{...}{additional parameters passed to load_ppi or the provided igraph method}
}
\value{
igraph
}
\description{
Method to filter graph based on any igraph method that scores verticies.
}
